KA-45_entityMOEXFeignClientVersion3
